Internship-report-PFE
This project is a comprehensive guide to securing information systems, covering both theoretical and practical aspects. The theoretical study in Chapter 2 examines the system of the company, its description, and the recommendations for securing it. The solutions proposed are also discussed in this chapter. Chapter 3 focuses on the practical application of the security measures discussed in Chapter 2. It covers the installation of the system, Active Directory Domain Services + DNS integration, folder and user creation, folder sharing configuration, protection against ransomware attacks, high availability of service, pfSense installation and configuration, security filtering using group policies, and data backup using Synology Drive and Windows Server.